beckend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cleanup==8.1.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-resized==1.0.2
djangorestframework==3.15.2
phonenumbers==8.13.45
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
    